% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.intcv.R
\name{dist.intcv}
\alias{dist.intcv}
\title{Built-in integration-based \%CV functions}
\usage{
dist.intcv(
  dist = "log",
  ...,
  exact = ifelse(dist == "log", TRUE, FALSE),
  lambda = NULL,
  fun = FALSE
)
}
\arguments{
\item{dist}{Selection of built-in distributions.}

\item{...}{passed to moment()}

\item{exact}{If there is an exact moment generating function, use that. Default TRUE only for log}

\item{lambda}{shape parameter for nonmemboxcox()}

\item{fun}{return function (for use in invcv())}
}
\value{
Percent CV
}
\description{
Built-in integration-based \%CV functions
}
