% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intcv.R
\name{intcv}
\alias{intcv}
\title{Integration-based CV\%}
\usage{
intcv(...)
}
\arguments{
\item{...}{Arguments passed to moment()}
}
\value{
Percent CV
}
\description{
Integration-based CV\%
}
\examples{
intcv(u = 1, v = 0.2, pdist = exp, qdist = log)
}
