% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invcv.R
\name{invcv}
\alias{invcv}
\title{Variance from CV\%}
\usage{
invcv(cvfun, cv, verbose = FALSE, ...)
}
\arguments{
\item{cvfun}{intcv()-based function}

\item{cv}{CV\% generated from cvfun}

\item{verbose}{extra output}

\item{...}{Other parameters to pass to cvfun}
}
\value{
Best-fit variance
}
\description{
Variance from CV\%
}
