% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moment_f.R
\name{moment_f}
\alias{moment_f}
\title{Integratable moment function}
\usage{
moment_f(x, u, v, n, pdist, qdist)
}
\arguments{
\item{x}{numeric vector}

\item{u}{mean}

\item{v}{variance}

\item{n}{moment number}

\item{pdist}{un-transform function for transformed random variable (eg, exp())}

\item{qdist}{transform function (eg, log())}
}
\value{
Point result of the moment function
}
\description{
Integratable moment function
}
