% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonmemboxcox.R
\name{nonmemboxcox}
\alias{nonmemboxcox}
\title{Box-Cox transform typically used in NONMEM}
\usage{
nonmemboxcox(x, lambda, inv = FALSE)
}
\arguments{
\item{x}{random vector}

\item{lambda}{shape parameter}

\item{inv}{inverse transform}
}
\value{
Box-Cox transformed or untransformed vector
}
\description{
NONMEM users will typically treat a Box-Cox-transformed \eta value as if it is log-normally transformed.
}
