% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numcv.R
\name{numcv}
\alias{numcv}
\title{Numeric CV\% of a sample}
\usage{
numcv(x, ...)
}
\arguments{
\item{x}{numeric vector}

\item{...}{other arguments for sd() and mean()}
}
\value{
Percent cv
}
\description{
Numeric CV\% of a sample
}
