% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.moment.R
\name{dist.moment}
\alias{dist.moment}
\title{Built-in moment functions}
\usage{
dist.moment(
  dist = "log",
  ...,
  exact = ifelse(dist == "log", TRUE, FALSE),
  lambda = NULL
)
}
\arguments{
\item{dist}{Selection of built-in distributions.}

\item{...}{passed to moment()}

\item{exact}{If there is an exact moment generating function, use that. Default TRUE only for log}

\item{lambda}{shape parameter for nonmemboxcox()}
}
\value{
moment
}
\description{
Built-in moment functions
}
